feat(grid): Add lazy coordinate generation
Refactor grid creation to support Dask-backed, lazy coordinates.

This commit addresses a key performance bottleneck by replacing eager `numpy.arange` calls with `dask.array.linspace` for coordinate generation. This ensures that the regridding pipeline remains lazy, significantly improving performance and reducing memory overhead for large datasets.

- Modified `create_lat_lon_coords` and `create_regridding_dataset` to accept a `use_dask` flag.
- Updated the `_create_source_grid_from_data` method in `CurvilinearRegridder` to use the new lazy generation for fallback coordinates.
- Added and updated tests to verify the new lazy generation functionality.
- Fixed linting issues to ensure code quality.

+++ src/monet_regrid/utils.py
@@ -120,8 +120,8 @@ def create_lat_lon_coords(grid: Grid, use_dask: bool = False) -> tuple[np.ndarra
     xp = da if use_dask else np
 
     # Calculate the number of points for latitude and longitude
-    n_lat = int(round((grid.north - grid.south) / grid.resolution_lat)) + 1
-    n_lon = int(round((grid.east - grid.west) / grid.resolution_lon)) + 1
+    n_lat = round((grid.north - grid.south) / grid.resolution_lat) + 1
+    n_lon = round((grid.east - grid.west) / grid.resolution_lon) + 1
 
     # Generate the coordinate arrays using linspace for robustness
     lat_coords = xp.linspace(grid.south, grid.north, n_lat)
